Post PR#79 cleanups: shared systemprune, per-name CreateShed lock, unified formatSize
Three follow-ups from PR #79's code review, grouped for one PR:

- Extract ~350 lines of duplicated DiskUsage/Prune helpers from
  internal/vz/system.go and internal/firecracker/system.go into a new
  leaf package internal/systemprune/. A bug fix now lands once instead
  of twice. Backend-specific pieces (ShedDiskEntry builders, instance
  candidate collection, VZ console-log truncation) stay local.
- Close the CreateShed/CopyRootfs TOCTOU: two concurrent shed create
  calls for the same name could both pass the metadata existence check,
  letting the second caller's os.Remove(dst) delete the first caller's
  just-cloned rootfs. A per-name sync.Mutex on Client serializes same-
  name creates while leaving different-name creates parallel. VZ and FC
  both get the lock; single-writer comments in rootfs.go now point at
  Client.acquireCreateLock as the enforcement point.
- Unify cmd/shed's two byte formatters: formatBytes (GB/MB only) is
  removed; its seven call sites in image.go now use formatSize, which
  moved from system.go to output.go. Sub-MB files in shed image list /
  prune now render as "0 B" / "512 B" instead of the misleading "0 MB".

Smoke-tested on macOS (VZ) and Linux amd64 (Firecracker): shed system df,
shed system prune --dry-run, single create, and concurrent same-name
create (one success, one SHED_ALREADY_EXISTS, rootfs intact at 20 GiB)
all behave as expected.
