+++ internal/vz/client_test.go
@@ -9,7 +9,9 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"sync"
 	"testing"
+	"time"
 
 	"github.com/charliek/shed/internal/config"
 	"github.com/charliek/shed/internal/vmutil"
@@ -282,6 +284,102 @@ func TestBuildCredentialShares(t *testing.T) {
 	}
 	if shareMap["cred-gh"] != "/home/user/.config/gh" {
 		t.Error("expected cred-gh share with source /home/user/.config/gh")
+	}
+}
+
+// TestAcquireCreateLock_SerializesSameName verifies that two acquireCreateLock
+// calls for the same shed name serialize — the second blocks until the
+// first releases. This is the lock that closes the CreateShed / CopyRootfs
+// TOCTOU race described in rootfs.go.
+func TestAcquireCreateLock_SerializesSameName(t *testing.T) {
+	c := &Client{}
+
+	release1 := c.acquireCreateLock("same")
+
+	acquired := make(chan struct{})
+	go func() {
+		release2 := c.acquireCreateLock("same")
+		close(acquired)
+		release2()
+	}()
+
+	select {
+	case <-acquired:
+		release1()
+		t.Fatal("second acquireCreateLock for same name should have blocked")
+	case <-time.After(100 * time.Millisecond):
+		// expected: still blocked on the first holder
+	}
+
+	release1()
+
+	select {
+	case <-acquired:
+		// expected: unblocked once the first released
+	case <-time.After(time.Second):
+		t.Fatal("second acquireCreateLock did not proceed after release")
+	}
+}
+
+// TestAcquireCreateLock_DifferentNamesDontBlock verifies that concurrent
+// creates for distinct names still run in parallel.
+func TestAcquireCreateLock_DifferentNamesDontBlock(t *testing.T) {
+	c := &Client{}
+
+	release1 := c.acquireCreateLock("a")
+	defer release1()
+
+	acquired := make(chan struct{})
+	go func() {
+		release2 := c.acquireCreateLock("b")
+		close(acquired)
+		release2()
+	}()
+
+	select {
+	case <-acquired:
+		// expected
+	case <-time.After(500 * time.Millisecond):
+		t.Fatal("acquireCreateLock for different names must not block")
+	}
+}
+
+// TestCreateShed_ConcurrentSameName_BlocksOnLock verifies the per-name
+// lock is actually wired into CreateShed. With the per-name mutex held
+// externally, a concurrent CreateShed for that name must block rather
+// than racing past the metadata existence check and corrupting the
+// first caller's rootfs via CopyRootfs's os.Remove(dst).
+func TestCreateShed_ConcurrentSameName_BlocksOnLock(t *testing.T) {
+	tmpDir := t.TempDir()
+	baseRootfs := filepath.Join(tmpDir, "base-rootfs.ext4")
+	if err := os.WriteFile(baseRootfs, []byte("rootfs"), 0644); err != nil {
+		t.Fatalf("write base rootfs: %v", err)
+	}
+
+	c := &Client{
+		cfg: &config.VZConfig{
+			BaseRootfs:  baseRootfs,
+			InstanceDir: tmpDir,
+		},
+		vms:         make(map[string]*VM),
+		createLocks: make(map[string]*sync.Mutex),
+		credMgr:     newTestCredMgr(),
+	}
+
+	release := c.acquireCreateLock("race-demo")
+	defer release()
+
+	done := make(chan struct{})
+	go func() {
+		_, _ = c.CreateShed(context.Background(), config.CreateShedRequest{Name: "race-demo"})
+		close(done)
+	}()
+
+	select {
+	case <-done:
+		t.Fatal("CreateShed must block on the per-name create lock while it is held")
+	case <-time.After(200 * time.Millisecond):
+		// expected: blocked on the lock.
 	}
 }
